Updated color logic to add convenience functions for converting (a, r, g, b) tuples to 0xAARRGGBB format. Added a greyscale color palette.

+++ evaluate.cc
@@ -2,5 +2,5 @@
 
 double linear_interpolation(Result *res, int max_iterations, double max_radius)
 {
-  return res->iterations / max_iterations;
+  return ((double)res->iterations) / ((double)max_iterations);
 }
